Split parsers into multiple files
